requirements.txt
pyyaml
numpy
pandas 
matplotlib